Conflicts:
	.settings/language.settings.xml
	appli/Logical/Logical.c
	appli/Logical/Logical.h
	appli/Logical/type.h
	appli/main.c

+++ appli/Abstract/Joystick/joystick.c
@@ -22,9 +22,9 @@ static volatile bool FLAG_1S = false;
 
 // private function
 	JOYSTICK_direction PRIVATE_JOYSTICK_getDirection(ADC_CHANNEL X, ADC_CHANNEL Y){
-		int16_t X_value = ADC_getValue(X)-MILIEU;
+		int16_t X_value = (int16_t)(ADC_getValue(X)-MILIEU);
 		//printf("X value : %d\n",MILIEU);
-		int16_t Y_value = ADC_getValue(Y)-MILIEU;
+		int16_t Y_value = (int16_t)(ADC_getValue(Y)-MILIEU);
 		printf("X value : %d\n",X_value);
 		printf("Y value : %d\n",Y_value);
 		if (Y_value<TOLERANCE_MIN && abs(Y_value)>abs(X_value)){
